Remove ReferenceIdentifierInterface
