+++ src/Identifier/ElementIdentifier.php
@@ -3,13 +3,11 @@
 namespace webignition\BasilModel\Identifier;
 
 use webignition\BasilModel\Value\LiteralValueInterface;
-use webignition\BasilModel\Value\ValueInterface;
 
-class ElementIdentifier extends AbstractIdentifier implements ElementIdentifierInterface
+class ElementIdentifier extends Identifier implements ElementIdentifierInterface
 {
     const DEFAULT_POSITION = 1;
 
-    private $value;
     private $position = 1;
 
     /**
@@ -19,17 +17,11 @@ class ElementIdentifier extends AbstractIdentifier implements ElementIdentifierI
 
     public function __construct(LiteralValueInterface $value, int $position = null, string $name = null)
     {
-        parent::__construct(IdentifierTypes::ELEMENT_SELECTOR, $name);
+        parent::__construct(IdentifierTypes::ELEMENT_SELECTOR, $value, $name);
 
         $position = $position ?? self::DEFAULT_POSITION;
 
-        $this->value = $value;
         $this->position = $position;
-    }
-
-    public function getValue(): ValueInterface
-    {
-        return $this->value;
     }
 
     public function getPosition(): int
@@ -50,19 +42,9 @@ class ElementIdentifier extends AbstractIdentifier implements ElementIdentifierI
         return $new;
     }
 
-    /**
-     * @param string $name
-     *
-     * @return IdentifierInterface|ElementIdentifierInterface
-     */
-    public function withName(string $name): IdentifierInterface
-    {
-        return parent::withName($name);
-    }
-
     public function __toString(): string
     {
-        $string = $this->value->getValue();
+        $string = parent::__toString();
 
         if ($this->parentIdentifier instanceof ElementIdentifierInterface) {
             $string = '{{ ' . $this->parentIdentifier->getName() . ' }} ' . $string;
