--- src/Identifier/ReferenceIdentifierInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace webignition\BasilModel\Identifier;
-
-interface ReferenceIdentifierInterface extends IdentifierInterface
-{
-
-}
